(FIX) Change hex functions to take a long instead of int. Also fix functionality of hex so it takes negatice numbers properly

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: oazlan <oazlan@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/31 12:58:13 by oazlan            #+#    #+#             */
-/*   Updated: 2025/12/31 15:42:21 by oazlan           ###   ########.fr       */
+/*   Updated: 2025/12/31 16:07:12 by oazlan           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,8 +29,8 @@ int		ft_intlen(int num);
 
 void	ft_printnum(int n);
 
-int		ft_hexlen(int num);
+int		ft_hexlen(long num);
 
-void	ft_print_hex(int n, char format_specifier);
+void	ft_print_hex(long n, char format_specifier);
 
 #endif
